backup: Write latest manifest file
Manifests currently cannot restore the "latest" backup because it would
require an expensive object-storage directory traversal. So instead it
defers to the pointer layout. The problem with this is that you then
loose manifest only features like setting the default branch properly.

Here we write two manifests: the normal backup manifest as before and an
additional "latest" manifest file. This latest manifest is overwritten
on each backup taken.

For WORM we would ideally not overwrite any files, but until this is
implemented we need something to fill the latest restore gap.

Changelog: changed