backup: Check for backup before deleting repo
Reorders the repository restore logic so that we only remove the
repository being restored if a backup actually exists for that
repository.

The previous order of operations caused issues when performing a full
restore of an instance that includes dangling repositories (repos
created after a backup was taken). Since it's a full restore, the
dangling repos were part of the inputs fed into `gitaly-backup`, and
thus were run through the `Restore` function of the `Manager`. This
function returns early if no backup was found, but the previous logic
removed repository before checking if a backup was available.