backup: Check for backup before deleting repo
Reorders the repository restore logic so that we only remove the
repository being restored if a backup actually exists for that
repository.

The previous order of operations caused issues when performing a full
restore of an instance that includes dangling repositories (repos
created after a backup was taken). Since it's a full restore, the
dangling repos were part of the inputs fed into `gitaly-backup`, and
thus were run through the `Restore` function of the `Manager`. This
function would remove the repo, and the caller of the function would
subsequently attempt to remove it again since it wasn't part of the
restore.

The "missing backup" test case has been updated to expect that the repo
exists. This is because cleanup of dangling repos is handled by the
caller of the `Restore` function.