Disable daily maintenance with transactions
The daily maintenance worker does not use a transaction and accesses
repositories directly. It thus conflicts with the work
TransactionManager is doing.

Changing daily maintenance to work with transactions is not worth it.
We need to pack the repositories actively as they are written to
since the file count impacts snapshotting efficiency.
TransactionManager sees all changes and thus can trigger the
maintenance when it is needed. Random daily walk is not useful in
this context as Gitaly no longer leaves garbage in the repositories.
The garbage left over by failures get cleaned up with the transactions.

This commit disables daily maintenance when transactions are enabled.
While at it, we only launch the daily maintenance worker if the feature
is actually enabled.