gitlab-ci: Add tests for reftable backend
The reftable backend is soon to be pushed upstream and might be released
in the next few git versions. Since the release is around the horizon,
it becomes ever so important that we test our repo against the new
backend to catch and fix bugs early in the lifecycle.

To support this, we add a new job to the GitLab CI to test against the
reftable backend. Since the reftable backend isn't officially merged to
next/master branches of git, we rely on the
`pks-reftables-initial-baseline` integration branch.

We do this by using the `GIT_DEFAULT_REF_FORMAT` env variable and
setting it to `reftable`, whenever the reftable tests are used.

Eventually we can get rid of the integration branch when the patch is
accepted upstream.