repository: Make GetInfoAttributes read from HEAD:.gitattribute
This is a workaround solution for N+1 problem we come across, see
https://gitlab.com/groups/gitlab-org/-/epics/9006#note_1730745750 for
details.

In order to remove the reliance of info/attribute and avoid breaking
Rails with TooManyInvocationsError. GetInfoAttributes is set to read
from HEAD:.gitattribute.

It will also try to delete info/attribute before reading.

After proper refactor on Rails to solve the N+1 problem in this issue
https://gitlab.com/gitlab-org/gitlab/-/issues/452425,
GetInfoAttributes can be removed totally.