ci: Don't inherit variables on downstream pipeline
The `inherit:variables` option is used to control whether variables set
in a parent CI configuration is inherited by its children. In our case,
the parent is the Gitaly CI pipeline and the child is the
`omnibus-gitlab-mirror` downstream pipeline, which we trigger here.

Disable variable inheritance as it appears to cause a compilation error
in the Omnibus pipeline when it attempts to build the Spamcheck binary.
Spamcheck is written in Go, so it's possible that we're setting some
specific toolchain options that conflict with its build process.

Other projects that trigger this downstream, such as `gitlab-shell`,
also explicitly disable variable inheritance. See
https://gitlab.com/gitlab-org/gitlab-shell/-/blob/03bb2b21172254b42ea5bbed0b6de9733b7dadf1/.gitlab-ci.yml#L195-196