gitlab-ci: Upgrade from debian bullseye to bookworm
The CI for gitaly runs on go v1.21. Since the minor version is not
defined, it uses go v1.21.3 when using debian:bullseye. This has a few
vulnerabilities [1]. While we do not dictate the minor versions for the
repository, by using the latest minor version, we can ensure that we are
not introducing those vulnerabilities through changes made to our code.

We do this by updating the container used by the CI, the latest
debian:bookworm uses go v1.21.5, which has the vulnerabilities fixed.
This requires changes to the expected postgresql schema.

The actual go version is handled by distribution and we only dictate the
minimum major version supported by gitaly.

[1]: https://go.dev/doc/devel/release#go1.21.0