Use physical logging for repository deletions
Repository deletions are currently handled by logging the logical
operation to delete the repository. Applying this is rather straight
forward by just calling `os.RemoveAll` on the directory and syncing
it. Nonetheless, this is extra handling that isn't needed as the
repository deletion could also be logged using the physical logging
protocol which logs the exact file system operations to perform. In
addition, `os.RemoveAll` could possibly remove more than was intended
if the repository contained unexpected files which were not present
when the log entry was built.

This commit moves from handling repository deletions logically to
logging the exact file system operations to perform the deletion.
This allows us to remove the redundant handling for deleting a
repository.

The logical operation is still left in the log entry as it is used
for conflict checks still. We'll later separate the conflict checking
from logging but first we want to migrate all logical operations to
physical operations.