Centralize fsync call before a transaction is committed
We have multiple places where we prepare files for logging. Each
preparation step is calling fsync independenly. We should add a more
general call before a transaction is committed. It would be less error
prone as we don't have to worry about syncing in each individual step
that prepares files for logging.