backup: Add SetHeadReference to Repository
Modifies the Repository interface with a new `SetHeadReference` method.
This method sets the symbolic HEAD reference of the repo; for local
repos it uses `SetDefaultBranch`, and for remote repos it uses the
`WriteRef` RPC which handles HEAD as a special case.

We update the HEAD separately from the rest of the refs, because the
associated branch is tracked by the manifest file rather than the .refs
artefact.