tests: Add and use `SkipWithReftable()`
Add a new function `SkipWithReftable()`, which can be used to skip a
test, if the reftable backend is enabled.

Let's use this and skip all failing tests. With this we can make the
reftable tests job required. This helps avoid creating new regressions
with respect to the reftable backend and gives clarity on what needs to
be fixed moving forward.