.gitlab-ci.yml: fix rails pipeline for security
The security repo also needs to ability to run rails tests to ensure we
haven't broken rails tests. To accomplish this, split out the rails
pipeline into a set of security specific ones. Note that for this to
work, a couple of CI/CD env vars need to be set on the
`gitlab-org/security/gitaly` project. Namely, GITLAB_PROJECT_TOKEN and
GITALY_REPO_DEPLOY_TOKEN.