cleanup: Allow empty requests for RewriteHistory
With 7630f4b496 (cleanup: Add RewriteHistory RPC, 2024-01-11) we added
validation that each request to `RewriteHistory` contain blobs and or
redactions. However, from the client perspective it is simpler to send
an initial request with just the repository, then stream in the
parameters in subsequent ones.

Update `RewriteHistory` to allow individual requests that contain
neither blobs or redactions. If no items are passed in any of the
client's request then we still return an `InvalidArgument` error.