backup: Separately handle skipped restores for latest and specific
Previously the manifest locator would fall back to pointer layout which
falls back to legacy layout. Legacy layout always returns a value on
find latest, it does not check to see if a backup actually exists.
Instead it relies on failing to find the bundle later in the process.

We cannot replicate this behaviour with manifest layout on its own. So
here we slightly change the error handling to match.