Remove packfile's static hashes from PackfilesState assertion
We have recently introduced support for repacking objects in the
write-ahead log. Most of the new tests use static hashing in packfile
assertion. It causes some troubles in some of Git's newest versions
because the hashing is non-deterministic, unstable, and subject to
change between Git versions. This commit modifies PackfilesState
asserter to avoid depending on packfiles' hashes and their orders.