Add database hook that is invoked after a log entry is deleted
Testing some functionality in the TransactionManager tests would
benefit from a database hook that is triggered after a log entry
is deleted from the database. This can be used for tests that want
to exercise scenarios where the log entry pruning is interrupted
after the protobuf message is deleted from the database but before
the files could be removed.