Prune files left over from interrupted log pruning
TransactionManager is currently not handling failures during log entry
pruning. It could be that the protobuf message of the log entry is
deleted from the database but the files are left on the disk. This
would lead to the files being left around indefinitely as we wouldn't
reattempt deleting them. This commit adds recovery logic for this
scenario. When initializing, we check whether there are stale files
in the WAL directory that do not belong to any log entry. If so, we
delete them, and proceed as normal. The stale files of deleted log
entries are now guaranteed to be removed as well.