Instrument TransactionManager with test hooks
TransactionManager's tests need to occasionally synchronize more
complex scenarios in tests. This is used to for example test that
if the manager crashes at a particular point, we can recover from
it gracefully.

The test hooks are currently plugged in transparently by wrapping
around the dependencies of the TransactionManager. For example, we
trigger some hooks when certain keys are read from the database.
This is complex to maintain as the internals keep changing. It's
also difficult to audit the code and see that the hooks are executed
in correct locations.

This commit replaces the dependency wrapping based test hooks with
direct instrumentation in TransactionManager. While generally it
would be preferable to keep testing specific code out of the production
code, in this case the cost is too high. In production, the test hooks
are simply empty function invocations which shouldn't add too much
overhead. As the hook invocations are also just function calls, they
don't clutter the production code too much either.

With the new approach, it's easy to insert hooks exactly where needed,
and see that they are invoked exactly where they should be.

The test cases themselves are mostly unchanged as the hooks generally
execute still at the same points. The only difference is the test case
testing initialization failures. The hook point is right before the
initialization finishes for clarity. Previously it was linked to a read
from the database that happened earlier in the initialization process
as we didn't have a good point to hook into.