Simplify transaction clean up
The TransactionManager is responsible for cleaning a transaction
after it is admitted. We previously needed to handle the scenario
where moving the files of a transaction to the WAL could succeed but
committing the protobuf message in the database fail. To handle that,
we had a mechanism to defer clean up of the files outside of the
transaction processing closure which returns the errors to the
transaction itself. Now that we commit the log entries only by moving
the directory in place, a partial commit is no longer a problem.

Simplify the transaction clean up to a simple defer.