Store log entries with lexicographically orderable names
TransactionManager is currently storing log entries in the
write-ahead log in a directory named by the log entry's LSN. The
string representation of the LSN is unpadded so the directory entries
are not lexicographically ordered in the same order as the actual LSN
values would be. Address this by padding the string representation of
an LSN with the full size of a uint64. This ensures their lexicographical
order is the same as their numerical order. Relying on this property,
we can simplify loading the log entries from the disk as they are
already returned in the correct order when reading the directory containing
them.