Add a helper package for building a write-ahead log entry
We're replacing the logical operations we currently log with physical
ones. This commits adds a helper package that can be used to build
the physical log entries.

The `Entry` type abstracts over a log entry. It records file system
operations that are to be performed in order to apply a transaction.
It also has a directory where it will stage the new files that are
introduced in a transaction.

Two helper methods are introduced that can be used to log some of
the existing operations we log:

- `RecordDirectoryCreation` logs a directory to be created along
  with its children when the log entry is applied. This can be used
  to for example log custom hook updates and repository creations.
- `RecordDirectoryDeletion` logs a directory to be removed with all
  its children. This can be used to log custom hook or repository
  removals.

The methods will be called when processing transactions to log the
physical changes they made. They both walk the directory hierarchies
and log the operations to be performed and stage the files needed.

A custom directory walker was added as the standard library ones do
not give enough control on walk ordering. To log creations, we need
to always log the directory to be created before its children. For
removals it is the reverse and the children should be removed first
and the parent directory last.

The package will be put to use in later commits.