Write custom hooks to disk in tests
TransactionManager's tests are currently not writing the custom hooks
to the disk. This will be a problem soon as we're about to log the
file changes to be done by reading them from the disk. This commit
starts writing them to the disk.

As we're now writing the hooks to the disk, we can't extract the
corrupt TAR in the test case asserting invalid hooks are rejected.
Let's remove the test case as we're about to remove the validation
functionality anyway in a later commit where we'll read the hooks
directly from the disk. The RPC handler writing the hooks would fail
if the TAR is invalid anyway.