Simplify transaction API for custom hooks
Custom hooks we're previously included in the log entry's manifest.
We thus captured the TAR that contains them and passed it around until
it finally gets logged. With the new physical logging approach, we'll
read the new state of custom hooks from the disk and log them directly
from there. As such, the API no longer needs to allow for passing a TAR.
`SetCustomHooks` is changed to `MarkCustomHooksUpdated` which simply hints
to the manager that custom hooks have been updated. The transaction manager
then knows to stage the custom hooks for committing. It checks the disk and
logs any possible new files, and deletes all of the old ones.

SetCustomHooks RPC was updated to no longer capture the TAR but rather
stream it directly for extraction.

The TransactionManager was previously verifying that the TAR that was given
to it was valid. This is no longer done as the hooks are taken directly from
the disk. The TransactionManager should focus on concurrency management and
not validate data. The RPC handlers can be expected to only stage correct
changes when they are working in isolation.