Use physical logging protocol for repository creations
Repository creations are currently logged by logging the logical
repository creation operation with its parameters. This leaves the
door open for minor behavior divergences. If Git is updated and
the logical creation is applied on a different version on one node
from others, the resulting state could be different. For example,
the new version could begin including some new default keys in the
configuration or some other files. Logging the repository's files
themselves instead of the creation operation would ensure that all
replicas will always look exactly the same.

This commit migrates repository creations to the physical logging
protocol. Instead of logging the command to create a repository,
we log the file system operations to create the repository. That way
every node applying the log entry will always end up with the same
result. Logging the file system operations is also more composable
approach than logging the logical operations. As long as the log
entry's producer knows which file system changes need to be performed,
all nodes will know how to apply the transaction regardless whether
they know about the logical operation that produced the log entry
yet. Most operations can be logged perfectly fine by just composing
the log entry from the few basic operations.

Logging the repository creation itself is fairly straightforward.
Since all of the state in a created repository is new, we'll walk
the repository and record the files and the directories that need
to be created.

While we generally log the repository exactly as it was produced by
the transaction, we override the object directory to only contain
the packs generated by the TransactionManager and the possible
alternate link. This ensures the object directory state is as the
TransactionManager expects, and only contains packed objects with
all their dependencies satisfied.

We still leave the logical operation in the LogEntry and its
associated code around. The conflict verification is still relying
on them. We'll later separate the conflict checking and the logging
protocol from each other.