panichandler: Log panic backtrace
Currently our panic handler logs the method and error message when an
RPC call panics. In many cases the error is sufficiently general that it
can be difficult to tell where in the call the panic occurred.

Add a `backtrace` field to the logged message to aid in researching the
cause of the panic.

Changelog: added