lint: Add 'blank' section to gci config
When importing a package without giving it a name, e.g. `import _ pkg`,
`goimports` will enforce a blank line above the enforced comment for the
import. However, the `gci` linter requires that all `default` imports be
contiguous.

To resolve this conflict, add a `blank` section to the `gci` config.
This will move all blank imports into their own section of the `import`
statement.