proto: Return error when path not treeish
Modifies the GetTreeEntries RPC so that an error is returned if the
provided path doesn't resolve to a treeish object, for example if a path
to a blob is provided. Previously, an empty result set would be returned
instead.

So that this works for the recursive case, a preliminary check is added
which calls `git cat-file -t` and verifies the resulting object type.
Without this check, the only error we get is `fatal: Not a valid object
name` as we pass both the resolved revision _and_ the path to
`git-ls-tree`. For example:

    > git ls-tree 5ba216ff70056be3e7de6f099ef00fa1b099563
    fatal: not a tree object

    > git ls-tree 5ba216ff70056be3e7de6f099ef00fa1b099563f:README.md
    fatal: Not a valid object name 5ba216ff70056be3e7de6f099ef00fa1b099563f:README.md

The change is made to `localrepo.ReadTree`, but the specific error is
only checked in the unified code path of the GetTreeEntries RPC, and
thus is only returned when the feature flag is enabled.