Use physical logging for default branch updates
This commit migrates default branch updates to the physical logging
protocol. Instead of logging the logical operation to update the
default branch, we log the exact file system operations to update
the default branch. This let's us remove the special handling for
updating HEAD and use the existing handling for applying the generic
file system operations.

The tests were updated to write the default branch update to the disk
where as they were previously staging it only in the transaction. This
matches what the RPC handlers would also do.

As we're now writing the default branch update using 'git symbolic-ref',
we remove the tests for invalid default branch updates. The command fails
when attempting to stage invalid default branch updates. The RPC handlers
should not stage bad state, and the transaction manager should focus on
resolving concurrency issues. We can assume the transactions run correctly
in isolation.

DefaultBranchUpdate is removed from the log entry as it is no longer
needed.